docs: adjust patterns docs with custom docs template and move resources into patterns folder
