+++ stories/patterns/docs/Template.mdx
@@ -1,0 +1,11 @@
+import { 
+  Description,
+  Meta, 
+  Stories, 
+  Title, 
+} from '@storybook/addon-docs/blocks';
+
+<Meta isTemplate />
+<Title />
+<Description />
+<Stories />
