+++ stories/patterns/examples/charts/BarChart.tsx
