+++ stories/patterns/examples/charts/MinimalLineChart.tsx
