+++ stories/patterns/examples/charts/MultiLineChart.tsx
