+++ stories/patterns/examples/charts/PieChart.tsx
